Introduce PropagateAndConvertNoAlias pass
This pass moves llvm.noalias intrinsics to the ptr_provenance path of
load/store instructions. At the same time, they are converted into
llvm.provenance.noalias intrinsics. For llvm.noalias intrinsics that are
returned or passed as arguments, a llvm.noalias.arg.guard is introduced,
combining the pointer path and the ptr_provenance path.

This migration allows other optimization passes to do changes on the
normal pointer path, without impacting the ptr_provenance path that
contains the restrict/noalias related information.

PropagateAndConvert: handle InsertElement (BADLY)

This is a temporary workaround to support a self build with
full LTO. For full support, a testcase with
vector if pointers must be created to show what should happen.