Introduce ConnectNoAliasDecl pass
This pass connects llvm.noalias.decl to llvm.noalias, llvm.provenance.noalias
and llvm.noalias.copy.guard intrinsics when those are associated with the
unknown function scope and depend on the same alloca object. The connection
allows a refinement of the noalias information for those intrinsics, resulting
in better code.